fix: Compile and lint the host crate on arm64 hosts
The crate carries a per-(arch, os) arm for almost everything, and the
AArch64 host column was simply missing.  `arch::trap::TrapFrame` fell
back to the placeholder unit struct on an AArch64 host while the real
AArch64 kernel modules were still compiled in, so their `frame.x5`
accesses failed; `validate_user_mapping`, the MMU facade, `fork` and the
process address-space accessors had no host arm either.

Widen the trap aliases to AArch64 hosts, add the missing host arms (a
state-less `PreparedProcessAddressSpace`, an `Unsupported` fork, a no-op
`validate_user_mapping`), and gate the x86_64-only test helpers to
x86_64.  AArch64 kernel code that has no host caller takes the
`cfg_attr(not(target_os = "none"), allow(dead_code))` the tree already
uses for that situation, along with the `# Safety` sections and the
clippy cleanups that the first lint of that code surfaced.

Verified with `clippy --all-targets -- -D warnings` for
aarch64-unknown-linux-gnu and aarch64-apple-darwin, with the three
bare-metal targets, the x86_64 hosts and the Linux test suite unchanged.

+++ src/arch/aarch64/smp.rs
@@ -3,6 +3,9 @@
 //! AArch64 SMP arch support: spin-table AP wakeup, MMU-config save/restore,
 //! GIC SGI (IPI) delivery, and AP entry-point logic.
 
+// AP bring-up and SGI delivery only run on bare metal.
+#![cfg_attr(not(target_os = "none"), allow(dead_code))]
+
 use crate::kernel::percpu::PerCpuData;
 use alloc::vec::Vec;
 use core::sync::atomic::AtomicU64;
@@ -248,7 +251,7 @@ fn discover_aps() -> Vec<(u32, u64)> {
     }
     let mut aps = Vec::new();
     for id in 1..total {
-        aps.push((id as u32, id as u64));
+        aps.push((id, id as u64));
     }
     crate::println!("[smp   ] FDT: {} CPUs total, {} AP(s)", total, aps.len());
     aps
